requirements.txt
Django>=5.2.5
psycopg2-binary>=2.9
python-dotenv>=1.0
gunicorn>=21.2
djangorestframework==3.16.1